Parameter Edit: Handle bitfields as ComboBoxes
This makes it possible to select more than one value
to be used in the parameter configuration, and since
the parameter is actually a bitfield - this is the
correct thing to do.

one of the drawbacks is that in the parameter listing
the value of the Fact will not be displayed as string
anymore, but with 'undefined: numerical-value', that is
the OR combination of the applied values.

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>