Clicking on the summary titles enter edit mode
Clicking on the sumarry titles enters the edit mode
On the tests with users, they always clicked on the
title (because it resembled a button or something) and
nothing happened. it was a bit confusing to have the
same information in two places and only one of them
being clickable.

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>