PX4: allow flight mode setup without sensors
Before this change it is not allowed to set up the flight modes unless
all sensor calibrations are completed.

I don't think that makes sense because it is possible to calibrate RC
without having the sensors calibrated, and it's the logical next step
right after that.

It is also a problem if HITL (hardware-in-the-loop) is used where the
sensors are not marked as calibrated but you still need to be able to
configure the flight modes. This is what prompted me to suggest this
change.