Remote id fronted: hide all indicators if RID not connected:
All this remote id feature depends on a device being detected
so it makes sense to not show the rest of the flags until
a healthy connection has been detected.

Also change label from RID COMMS to NOT CONNECTED when there is
no RID connection, so it is more descriptive to the user.